chore(wave-cleanup): close 6 deferred cycles (#431)
* chore(deps): break award + campaign-context-panel cycles

Pattern 4: extract shared types to *.types.ts leaf files.

- src/components/award/awardIcons.tsx <-> awardIconComponents.tsx
  IconProps moved to awardIcons.types.ts; both halves import the leaf.
- src/components/campaign/ContextPanel.tsx <-> ContextPanelTabs.tsx
  ContextPanelMode + 5 data interfaces moved to ContextPanel.types.ts;
  re-exported from ContextPanel.tsx so existing public API is preserved
  (storybook + other consumers unaffected).

Verified: madge -> 0 circular deps in src/components/award and
src/components/campaign.

* chore(deps): break repair-bay + unit-card-expanded cycles

Pattern 4 + direct re-route.

- src/components/gameplay/pages/repair/bay/RepairBayPage.RepairWorkspace.tsx
  Imported UnitRepairCard from RepairBayPage.sections (which only re-exports
  it from RepairBayPage.UnitRepairCard). Switched to direct import from
  RepairBayPage.UnitRepairCard.tsx so RepairWorkspace no longer depends on
  the orchestrator that imports it back.
- src/components/unit-card/UnitCardExpanded.tsx <-> UnitCardExpandedSections.tsx
  EquipmentEntry, CriticalSlotSummary, UnitCardExpandedProps moved to
  UnitCardExpanded.types.ts; both halves import the leaf and the
  orchestrator re-exports the types so unit-card/index.ts public API is
  preserved.

Verified: madge -> 0 circular deps in src/components/gameplay/pages/repair/bay
and src/components/unit-card.

* chore(deps): break BlkExportServiceCore <-> exporters cycle

Extracts IBlkExportResult and ExportableUnitState union into a new
BlkExportServiceCore.types.ts leaf. Both BlkExportServiceCore.ts and
BlkExportServiceCore.exporters.ts import from the leaf; the orchestrator
re-exports for backward compat with any external callers.

Pattern 1 mechanical fix (same as PRs #419, #420, #422). 5th cycle of
the 6 deferred Pattern 4 cycles closed.

* chore(deps): break aerospace validationRules 3-way cycle

Closes the 3-way circular dependency between
utils/construction/aerospace/{validationRules, bombBays,
wingHeavyWeapons}.ts. Extracts AerospaceValidationError to a new
validationRules.types.ts leaf; bombBays and wingHeavyWeapons import
from the leaf instead of from validationRules.ts. validationRules.ts
re-exports the type for backward compat.

Last of the 6 deferred Pattern 4 cycles closed (with
BlkExportServiceCore from c241c83e).

madge after: 0 cycles in services/conversion + utils/construction/aerospace + components.
