+++ src/utils/construction/aerospace/wingHeavyWeapons.ts
@@ -23,7 +23,7 @@
  * Requirement: Equipment Mounting per Arc
  */
 
-import type { AerospaceValidationError } from './validationRules';
+import type { AerospaceValidationError } from './validationRules.types';
 
 import {
   AerospaceArc,
